requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.8.0
django-extra-fields==2.0.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
Pillow==6.2.1
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.5
requests==2.22.0
six==1.13.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
whitenoise==5.2.0
xmltodict==0.12.0
